Define image and active storage tables
This commit adds migrations for active storage and image.

The image model acts as a wrapper to the file stored with active storage.

**NB:**

Due to the shortcomings with SQLITE data types.. an after_initialize callback
has been used to work around this.

This isn't ideal and has been taken due to the non production nature of this
repo.